fix: correct routing for API
