+++ firebase.json
@@ -1,9 +1,10 @@
 {
   "hosting": {
+    "public": "public",
     "rewrites": [
       {
-        "source": "/v1/**",
-        "function": "v1"
+        "source": "**",
+        "function": "api"
       }
     ]
   },
